- File size information is now parsed and displayed for NyaaTorrents feed. - Statistics dialog now displays the amount of scores next to score distribution bars. - Edit button in Now Playing dialog is now shown as "Add to list" for items not in list. - When clicked on Now Playing balloon tooltip, Now Playing dialog is shown and main window is activated. - Modified the priority constant of GetMostCommonCharacter() a bit so that ".Ro-Kyu-Bu!.SS.01." is recognized correctly. - Fixed a bug where percent character in file names (e.g. "Magical☆Star Kanon 100% OVA") caused ReplaceVariables() to fail. - Fixed a bug where selected list item's progress buttons were still drawn after switching tabs. - Fixed a bug where removing fansub group preference from anime information dialog didn't work. - Fixed a bug where dropping folders over root folders page didn't work.
git-svn-id: http://taiga.googlecode.com/svn/trunk@216 77add795-4bdf-f0f6-a574-f4d068491005

+++ anime_db.cpp
@@ -598,7 +598,7 @@ bool SeasonDatabase::IsRefreshRequired() {
   int count = 0;
   bool required = false;
 
-  for (int i = 0; i < items.size(); i++) {
+  for (size_t i = 0; i < items.size(); i++) {
     int anime_id = items.at(i);
     auto anime_item = AnimeDatabase.FindItem(anime_id);
     if (anime_item) {
@@ -620,7 +620,7 @@ void SeasonDatabase::Review(bool hide_hentai) {
   mal::GetSeasonInterval(name, date_start, date_end);
 
   // Check for invalid items
-  for (int i = 0; i < items.size(); i++) {
+  for (size_t i = 0; i < items.size(); i++) {
     bool invalid = false;
     int anime_id = items.at(i);
     auto anime_item = AnimeDatabase.FindItem(anime_id);
